Mark deleted maintenance in schedule map

+++ src/main/java/org/acme/foodpackaging/domain/PackagingSchedule.java
@@ -16,7 +16,7 @@ import ai.timefold.solver.core.api.solver.SolverStatus;
 import lombok.Getter;
 import lombok.Setter;
 import org.acme.foodpackaging.record.DbJobRow;
-import org.acme.foodpackaging.record.DbMaintenanceRow;
+import org.acme.foodpackaging.dto.DbMaintenanceRow;
 
 @Setter
 @Getter
@@ -38,7 +38,7 @@ public class PackagingSchedule {
 
     private Map<Integer, DbJobRow> dbJobRowMap;
 
-    private Map<Integer, DbMaintenanceRow> dbMaintenanceRowMap;
+    private Map<Long, DbMaintenanceRow> dbMaintenanceRowMap;
 
     private  Map<Integer, Job> jobIdMap;
